Fix not working validation
Since 4fb0f1724c460f019a0500201c9247d49f49a51e the validation shows errors for hidden fields in my forms. This is caused by a not reset $fld variable in the loop, when $formats[$name] is not set. In that case the old $fld value was used with the value of the new field.